[BE] refactor : 스프링 이벤트 위치 변경
