+++ backend/src/main/java/moim_today/application/admin/member/AdminMemberService.java
@@ -3,9 +3,8 @@ package moim_today.application.admin.member;
 import moim_today.dto.member.MemberResponse;
 import moim_today.global.error.ForbiddenException;
 import moim_today.persistence.entity.member.MemberJpaEntity;
-import moim_today.persistence.repository.meeting.joined_meeting.JoinedMeetingRepository;
 import moim_today.persistence.repository.member.MemberRepository;
-import moim_today.spring_event.event.AdminMemberDeleteEvent;
+import moim_today.global.spring_event.event.AdminMemberDeleteEvent;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
